apps/backend/requirements.txt
flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp
opentelemetry-instrumentation-dbapi
mysql-connector-python